// src/types/types.ts
export interface FallingHeart {
  id: number;
  left: number;
  delay: number;
  duration: number;
  size: number;
}

export interface ExplosionHeart {
  id: number;
  angle: number;
  distance: number;
  duration: number;
  delay: number;
}

export interface ButtonPosition {
  top: string;
  left: string;
}